move cta card to discrete file
