+++ src/scenes/Home/TerrawareHomeView.tsx
@@ -2,9 +2,8 @@ import React, { useEffect, useMemo, useState } from 'react';
 import { useMixpanel } from 'react-mixpanel-browser';
 import { useNavigate } from 'react-router-dom';
 
-import { Box, Container, Grid, SxProps, Typography, useTheme } from '@mui/material';
+import { Box, Container, Grid, Typography, useTheme } from '@mui/material';
 import { Icon, IconName } from '@terraware/web-components';
-import { Props as ButtonProps } from '@terraware/web-components/components/Button/Button';
 import { getDateDisplayValue, useDeviceInfo } from '@terraware/web-components/utils';
 
 import PageHeader from 'src/components/PageHeader';
@@ -12,7 +11,6 @@ import Link from 'src/components/common/Link';
 import PageCard from 'src/components/common/PageCard';
 import PlantingSiteSelector from 'src/components/common/PlantingSiteSelector';
 import TfMain from 'src/components/common/TfMain';
-import Button from 'src/components/common/button/Button';
 import {
   ACCELERATOR_LINK,
   APP_PATHS,
@@ -36,6 +34,7 @@ import { isAdmin } from 'src/utils/organization';
 import useMapboxToken from 'src/utils/useMapboxToken';
 import { useDefaultTimeZone } from 'src/utils/useTimeZoneUtils';
 
+import CTACard from './TerrawareHomeView/CTACard';
 import OrganizationStatsCard, { OrganizationStatsCardRow } from './TerrawareHomeView/OrganizationStatsCard';
 
 export type StatsCardItemProps = {
@@ -293,117 +292,6 @@ export const PlantingSiteStats = () => {
       >
         {token && (
           <img alt='Mapbox Static Map with Boundaries' src={staticMapURL} style={{ width: '100%', height: 'auto' }} />
-        )}
-      </Box>
-    </Box>
-  );
-};
-
-type CTACardProps = {
-  buttonsContainerSx?: SxProps;
-  description: string | (string | JSX.Element)[];
-  imageSource?: string;
-  padding?: number | string;
-  primaryButtonProps?: ButtonProps;
-  secondaryButtonProps?: ButtonProps;
-  title?: string | (string | JSX.Element)[];
-};
-
-const CTACard = ({
-  buttonsContainerSx,
-  description,
-  imageSource,
-  padding = '24px',
-  primaryButtonProps,
-  secondaryButtonProps,
-  title,
-}: CTACardProps): JSX.Element => {
-  const { isDesktop, isMobile, isTablet } = useDeviceInfo();
-  const theme = useTheme();
-
-  return (
-    <Box
-      sx={{
-        alignItems: 'center',
-        background: theme.palette.TwClrBg,
-        borderRadius: '8px',
-        display: 'flex',
-        flexDirection: isDesktop ? 'row' : 'column',
-        height: '100%',
-        justifyContent: 'space-between',
-        padding,
-      }}
-    >
-      <Box
-        sx={{
-          display: 'flex',
-          flexDirection: isMobile ? 'column' : 'row',
-        }}
-      >
-        {imageSource && (
-          <Box
-            sx={{
-              marginBottom: isMobile ? '32px' : 0,
-              marginRight: isMobile ? 0 : '32px',
-              textAlign: 'center',
-            }}
-          >
-            <img src={imageSource} />
-          </Box>
-        )}
-        <Box>
-          {title && (
-            <Typography
-              component='p'
-              variant='h6'
-              sx={{
-                color: theme.palette.TwClrTxt,
-                fontSize: '16px',
-                fontWeight: 600,
-                lineHeight: '24px',
-              }}
-            >
-              {title}
-            </Typography>
-          )}
-          <Typography
-            component='p'
-            variant='h6'
-            sx={{
-              color: theme.palette.TwClrTxt,
-              fontSize: '16px',
-              fontWeight: 400,
-              lineHeight: '24px',
-            }}
-          >
-            {description}
-          </Typography>
-        </Box>
-      </Box>
-
-      <Box
-        sx={[
-          {
-            display: 'flex',
-            flexDirection: isMobile ? 'column' : 'row',
-            marginLeft: isDesktop ? '27px' : 0,
-            marginTop: isMobile || isTablet ? '32px' : 0,
-            whiteSpace: 'nowrap',
-          },
-          ...(Array.isArray(buttonsContainerSx) ? buttonsContainerSx : [buttonsContainerSx]),
-        ]}
-      >
-        {primaryButtonProps && <Button priority='secondary' type='productive' {...primaryButtonProps} />}
-        {secondaryButtonProps && (
-          <Button
-            priority='secondary'
-            style={{
-              marginLeft: isMobile ? 0 : '19px',
-              marginTop: isMobile ? '19px' : 0,
-            }}
-            type='passive'
-            {...secondaryButtonProps}
-          />
         )}
       </Box>
     </Box>
